pyproject.toml
[project]
name = "deeplearning-mcp-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.53.0",
    "arxiv>=2.2.0",
    "mcp>=1.9.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]
